Fix DeleteThumbs crash when thumbs directory does not exist
Guard against DirectoryNotFoundException when wwwroot/thumbs has not
yet been created (fresh install or different deployment path).

+++ Nop.Plugin.Misc.Watermark/Services/MiscWatermarkPictureService.cs
@@ -97,6 +97,8 @@ namespace Nop.Plugin.Misc.Watermark.Services
         {
             var defaultThumbsPath = _fileProvider.GetAbsolutePath(NopMediaDefaults.ImageThumbsPath);
             var imageDirectoryInfo = new DirectoryInfo(defaultThumbsPath);
+            if (!imageDirectoryInfo.Exists)
+                return Task.CompletedTask;
             foreach (var fileInfo in imageDirectoryInfo.GetFiles())
                 fileInfo.Delete();
